+++ src/main/scala/org/podval/tools/test/framework/FrameworkProvider.scala
@@ -1,42 +1,45 @@
 package org.podval.tools.test.framework
 
+import org.slf4j.{Logger, LoggerFactory}
 import sbt.testing.{Framework, Runner}
 
 sealed abstract class FrameworkProvider:
-  def frameworkName: String
+  def frameworkName: String = framework.name
+  def frameworkDescriptor: FrameworkDescriptor = FrameworkDescriptor.forName(frameworkName)
+  lazy val framework: Framework = frameworkOpt.get
 
-  def framework: Framework
-  
-  final def frameworkDescriptor: FrameworkDescriptor = FrameworkDescriptor.forName(frameworkName)
-  
-  final def makeRunner(
+  final def frameworkOpt: Option[Framework] =
+    try Class
+      .forName(frameworkDescriptor.className)
+      .getDeclaredConstructor()
+      .newInstance() match
+      case framework: Framework => Some(framework)
+      case other =>
+        FrameworkProvider.logger.error(s"${other.getClass.getName} is not an SBT framework!")
+        None
+    catch
+      case _: ClassNotFoundException => None
+
+  final def runner(
     includeTags: Array[String],
     excludeTags: Array[String]
-  ): Runner =
-    val args: Array[String] = frameworkDescriptor.args(
+  ): Runner = framework.runner(
+    frameworkDescriptor.args(
       includeTags = includeTags,
       excludeTags = excludeTags
-    )
+    ),
+    Array.empty,
+    framework.getClass.getClassLoader
+  )
 
-    // We are running the runner in *this* JVM, so remote arguments are not used?
-    val remoteArgs: Array[String] = Array.empty
-
-    val frameworkClassLoader: ClassLoader = framework.getClass.getClassLoader
-
-    framework.runner(
-      args,
-      remoteArgs,
-      frameworkClassLoader
-    )
-    
 object FrameworkProvider:
-  final class Forking(
-    override val frameworkName: String
-  ) extends FrameworkProvider:
-    override lazy val framework: Framework = frameworkDescriptor.newInstance.get
+  private val logger: Logger = LoggerFactory.getLogger(FrameworkDescriptor.getClass)
 
-  final class NonForking(
-    override val framework: Framework
-  ) extends FrameworkProvider:
-    override def frameworkName: String = framework.name
-  
+  def apply(name: String): FrameworkProvider = new FrameworkProvider:
+    override def frameworkName: String = name
+
+  def apply(it: Framework): FrameworkProvider = new FrameworkProvider:
+    override lazy val framework: Framework = it
+
+  def apply(descriptor: FrameworkDescriptor): FrameworkProvider = new FrameworkProvider:
+    override def frameworkDescriptor: FrameworkDescriptor = descriptor